quick_batch/controller_app/requirements.txt
Flask
requests
pyyaml
docker